+++ dvr_scan/config.py
@@ -369,6 +369,9 @@ CONFIG_MAP: ConfigDict = {
     # Logging
     "verbosity": "info",
     "save-log": True,
+    # TODO: max-log-size is not implemented. This is kept for backwards compatibility with older
+    # config files. Re-implementing size-based rotation is likely overly complicated since log
+    # files are not expected to be very large.
     "max-log-size": 20000,
     "max-log-files": 4,
     # Development