+++ dvr_scan/controller.py
@@ -28,7 +28,7 @@ from dvr_scan.shared import ScanSettings, init_logging, init_scanner, logfile_pa
 
 logger = logging.getLogger("dvr_scan")
 
-LOGFILE_PATH = logfile_path(logfile_name="dvr-scan.log")
+LOGFILE_PATH = logfile_path(name_prefix="dvr-scan")
 
 
 def _preprocess_args(args):
@@ -101,8 +101,8 @@ def parse_settings() -> ty.Optional[ScanSettings]:
         if config.get("save-log"):
             setup_logger(
                 logfile_path=LOGFILE_PATH,
-                max_size_bytes=config.get("max-log-size"),
                 max_files=config.get("max-log-files"),
+                name_prefix="dvr-scan",
             )
         failed_to_load_config = False
     except ConfigLoadFailure as ex: